requirements.txt
google-cloud-aiplatform
google-adk
litellm
mcp
pydantic
pydantic-settings
pydantic-ai
instructor
langchain
langgraph
llama-index
tenacity
json-repair
playwright
pygithub
python-dotenv
pytest
rich
pytest-asyncio
prompt_toolkit
duckduckgo-search
